#36 - dependencies core refactor
