+++ tests/OpenApiCompatibilityTest.php
@@ -6,72 +6,27 @@ namespace Blumilk\OpenApiToolbox\Tests;
 
 use Blumilk\OpenApiToolbox\Config\DocumentationConfig;
 use Blumilk\OpenApiToolbox\Config\Format;
-use Blumilk\OpenApiToolbox\OpenApiSpecification\SpecificationBuilder;
-use KrzysztofRewak\OpenApiMerge\Writer\Exception\InvalidFileTypeException;
-use PHPUnit\Framework\Assert;
-use PHPUnit\Framework\TestCase;
+use Blumilk\OpenApiToolbox\OpenApiCompatibility\OpenApiCompatibility;
+use Orchestra\Testbench\TestCase;
 
 class OpenApiCompatibilityTest extends TestCase
 {
-    /**
-     * @throws InvalidFileTypeException
-     */
-    public function testSingleFileDocumentationBuild(): void
+    use OpenApiCompatibility;
+
+    public function getDocumentationConfig(): DocumentationConfig
     {
-        $config = new DocumentationConfig([
+        return new DocumentationConfig([
             "specification" => [
                 "index" => "openapi.yml",
-                "path" => realpath(__DIR__ . "/mocks/yml/singleFileDocumentation"),
+                "path" => realpath(__DIR__ . "/mocks/yml/loginDocumentation"),
                 "allow_multiple_files" => false,
             ],
-            "format" => Format::Yml,
+            "format" => Format::Yaml,
         ]);
-
-        $builder = new SpecificationBuilder($config);
-        $result = $builder->build();
-
-        Assert::assertStringContainsString("API Documentation", $result);
     }
 
-    /**
-     * @throws InvalidFileTypeException
-     */
-    public function testMultipleFilesDocumentationWithSingleFileConfigurationBuild(): void
+    public function testDocumentationCompatibility(): void
     {
-        $config = new DocumentationConfig([
-            "specification" => [
-                "index" => "openapi.yml",
-                "path" => realpath(__DIR__ . "/mocks/yml/multipleFilesDocumentation"),
-                "allow_multiple_files" => false,
-            ],
-            "format" => Format::Yml,
-        ]);
-
-        $builder = new SpecificationBuilder($config);
-        $result = $builder->build();
-
-        Assert::assertStringContainsString("API Documentation", $result);
-        Assert::assertStringNotContainsString("Ping endpoint", $result);
-    }
-
-    /**
-     * @throws InvalidFileTypeException
-     */
-    public function testMultipleFilesDocumentationBuild(): void
-    {
-        $config = new DocumentationConfig([
-            "specification" => [
-                "index" => "openapi.yml",
-                "path" => realpath(__DIR__ . "/mocks/yml/multipleFilesDocumentation"),
-                "allow_multiple_files" => true,
-            ],
-            "format" => Format::Yml,
-        ]);
-
-        $builder = new SpecificationBuilder($config);
-        $result = $builder->build();
-
-        Assert::assertStringContainsString("API Documentation", $result);
-        Assert::assertStringContainsString("Ping endpoint", $result);
+        $this->postJson("login");
     }
 }
